i9300: audio: fix low latency audio support
We are advertising low latency audio support but we were not setting
the proper buffer sizes and audio latencies for low latency audio.
This caused issues with choppy and stuttery audio when the application
sets the low latency audio flag.

This patch fixes choppy sound issues with Games built with Unity3D 5.2
(Temple Run, Subway Surfers)

http://forum.unity3d.com/threads/android-sound-problem.359341/

Change-Id: I46b4786aa9c3a65a9820c127adb443959a28eb15